+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 Changelog
 =========
 
-1.0.1:
-------
-- switched build system to electron-builder
+- 1.0.2
+  - prompt user for Spotify App client id and secret instead of reading it from *config.json*
+- 1.0.1
+  - switched build system to electron-builder